requirements.txt
torchvision
tqdm
pandas